fix: setting storybook for remix vite
